+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+## x.x.x (Released 2022-xx-xx)
+
+- Individual validation issues can be set as ignored.
+- Minor improvements to error handling.
+
 ## 0.9.1 (Released 2022-01-04)
 
 - If auto-save is enabled, a save prompt will be displayed when first adding an arrangement to a new project.